refactor: extract query handlers from tools.ts to query.ts
